requirements.txt
catboost==1.2.1.1
fastparquet==2023.8.0
fonttools==4.42.1
jupyter 
matplotlib==3.7.2
notebook
numpy==1.24.4
onnxmltools==1.11.2
openpyxl==3.1.2
pandas==2.0.3
pandocfilters==1.5.0
parquet==1.3.1
pydot==1.4.2
PyYAML==6.0.1
requests==2.31.0
scikit-image==0.21.0
scikit-learn==1.3.0
scipy==1.10.1
tqdm==4.66.1
urllib3==2.0.4
xgboost==2.0.0
python-dateutil==2.8.2
lightgbm==3.3.5
skl2onnx==1.15.0
onnx==1.14.1
onnxruntime==1.16.1
packaging==23.1
easydict==1.10